Move waitForBackend for util.go
